backend/requirements.txt
fastapi
uvicorn
email_validator
google-genai
openai
python-dotenv
